requirements.txt
# Core
Django>=6.0
django-environ>=0.11.2
whitenoise>=6.7
psycopg2-binary>=2.9
pillow>=10.2.0

# Development & Testing
pytest>=8.0
pytest-django>=4.8

# Code Quality
ruff>=0.4
black>=24.0
mypy>=1.10
django-stubs>=5.0
